Move and refactor grid diff + precond tests
